# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BanyuHub.space API Documentation",
        "description": "Layanan API untuk sistem informasi event dan komunitas akademik lintas kampus di wilayah Banyumas yang mendukung autentikasi pengguna, manajemen event, registrasi peserta, ulasan event, serta integrasi aplikasi web dan mobile.\n\nDokumentasi API ini dirancang menggunakan standar REST API untuk memudahkan proses pengembangan dan integrasi sistem pada aplikasi mobile Flutter maupun aplikasi web Laravel.",
        "contact": {
            "name": "Tim Developer BanyuHub",
            "email": "developer@banyuhub.space"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Primary API Server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register pengguna baru",
                "description": "Mendaftarkan pengguna (mahasiswa/civitas) baru ke sistem.\n\nProses ini memvalidasi input data nama, email unik, dan password minimal 6 karakter.\nSetelah data disimpan dengan password terenkripsi, sistem langsung menerbitkan\ntoken akses API (Sanctum) agar pengguna baru bisa langsung terautentikasi.",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login pengguna",
                "description": "Melakukan proses login pengguna.\n\nMemverifikasi kecocokan email dan password yang dikirimkan.\nJika sukses, token akses API (Sanctum) baru akan diterbitkan untuk sesi pengguna ini.\nJika gagal, sistem akan mengembalikan pengecualian validasi berupa pesan kesalahan kredensial.",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authenticated user",
                "description": "Mengambil informasi profil pengguna yang sedang login.\n\nEndpoint ini memerlukan token akses Bearer yang valid. Data yang dikembalikan\nberupa objek detail profil pengguna (ID, nama, email, role, dan penanda verifikasi).",
                "operationId": "e44b931999dcfc054d7a30b300537f43",
                "responses": {
                    "200": {
                        "description": "Authenticated user data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout pengguna",
                "description": "Melakukan proses logout pengguna.\n\nMencabut dan menghapus token akses API (Sanctum) yang sedang aktif digunakan,\nsehingga sesi autentikasi pengguna tersebut diakhiri dan tidak bisa lagi mengakses endpoint terproteksi.",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Logout successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Dapatkan semua event dengan fitur pencarian dan filter",
                "description": "Mengambil daftar event dengan dukungan pencarian, filter status, dan paginasi.\n\nMemungkinkan client mencari event berdasarkan nama, deskripsi, atau lokasi,\nserta melakukan filter berdasarkan status ('Mendatang', 'Sedang Berjalan', 'Selesai', 'Dibatalkan').\nHasil pencarian dibatasi menggunakan paginasi default sebanyak 10 item per halaman.",
                "operationId": "e529c81e6aaec82f49950a87ed881f57",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari berdasarkan nama, deskripsi, atau lokasi",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter berdasarkan status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Mendatang",
                                "Sedang Berjalan",
                                "Selesai",
                                "Dibatalkan"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Dapatkan detail event",
                "description": "Mengambil detail lengkap suatu event berdasarkan ID.\n\nMencari event yang sesuai dengan ID yang diberikan. Jika data ditemukan,\nakan dikembalikan beserta respons sukses. Jika tidak ditemukan, mengembalikan status error 404.",
                "operationId": "c0f602d773798994c228d99dbd63b842",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID event yang dicari.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/register-event": {
            "post": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Daftar ke sebuah event",
                "description": "Mendaftarkan pengguna yang sedang login ke sebuah event.\n\nMelakukan validasi ketat sebelum pendaftaran dibuat:\n1. Memastikan event ada di database.\n2. Memastikan status event bukan 'Selesai' atau 'Dibatalkan'.\n3. Memastikan kuota peserta event masih tersedia (lebih besar dari 0).\n4. Memastikan pengguna belum pernah mendaftar di event yang sama sebelumnya (mencegah duplikasi).\n\nJika semua validasi lolos, pendaftaran disimpan dengan status awal 'Pending' (dan memicu\nmodel observer untuk otomatis mengurangi kuota event sebanyak 1).",
                "operationId": "692fbcdfcf936f3a3a0b78d0550a489d",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Registered successfully"
                    },
                    "400": {
                        "description": "Kuota penuh atau sudah terdaftar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-registrations": {
            "get": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Dapatkan status registrasi event pengguna",
                "description": "Mengambil daftar riwayat pendaftaran event milik pengguna yang sedang login.\n\nMengembalikan seluruh record pendaftaran yang dibuat oleh pengguna, lengkap dengan\nrelasi data detail event-nya, diurutkan dari pendaftaran yang terbaru.",
                "operationId": "8eee03fe19237fe180f869f27278df21",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/events/{event}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Dapatkan semua ulasan untuk event tertentu",
                "description": "Mengambil seluruh daftar ulasan untuk event tertentu.\n\nMengembalikan ulasan beserta nama pengguna yang memberikan rating dan komentar.",
                "operationId": "94d5427dc3a2a94dbb2ad60e1d101d88",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Objek event yang diikat (Route Model Binding).",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan ulasan"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Tambahkan ulasan ke event",
                "description": "Mengirimkan ulasan (rating dan komentar) baru untuk suatu event.\n\nMelakukan pengecekan aturan bisnis:\n- Memastikan status event sudah 'Selesai'.\n- Memastikan user terdaftar dengan status RSVP 'Approved'.\n- Mencegah duplikasi ulasan dari user yang sama.\n- Memvalidasi rating harus berupa angka bulat 1-5, dan komentar maksimal 1000 karakter.",
                "operationId": "2b4652ff9813984267b83715620e4e1e",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Objek event yang diulas (Route Model Binding).",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Bagus banget!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ulasan berhasil ditambahkan"
                    },
                    "400": {
                        "description": "Event belum selesai"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    },
                    "403": {
                        "description": "Tidak memiliki akses untuk mengulas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan Sanctum Token yang didapat dari endpoint Login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Manajemen autentikasi mahasiswa (Login, Register, Logout)."
        },
        {
            "name": "Events",
            "description": "Eksplorasi, pencarian, dan detail kegiatan atau organisasi kampus se-Banyumas."
        },
        {
            "name": "Registrations",
            "description": "Manajemen partisipasi mahasiswa (pendaftaran event dan riwayat RSVP)."
        },
        {
            "name": "Reviews",
            "description": "Ulasan dan sistem rating komunitas terhadap event yang sudah selesai."
        },
        {
            "name": "User",
            "description": "User"
        }
    ]
}